refactor: split doc generators and share common parts (#16)
문서를 둘로 나눈다 — design-system.html은 토큰 값만, components.html은 셸·컴포넌트
렌더. 읽는 목적이 다르다: 전자는 "쓸 값을 고르러" 오고 후자는 "이 화면과 비슷한 걸
어떻게 짰나"를 보러 온다. 한 파일에 있으면 둘 다 찾기 어려워진다.

토큰 읽기와 문서 뼈대(헤더·섹션·details)를 doc-shared.mjs로 뺐다. 두 생성기가
각자 베끼면 스타일이 갈라지고, 그러면 "같은 시스템의 두 문서"처럼 보이지 않는다.

동작은 바뀌지 않는다 — design-system.html의 산출물은 컴포넌트 절이 빠진 것 외에 같다.
